added color extension
