+++ catalog/gallery/lib/catalog/catalog_app_colors_screen.dart
@@ -56,15 +56,15 @@ extension _CatalogScreenExtensions on _CatalogColors {
       case _CatalogColors.secondary:
         return context.theme.colors.secondary;
       case _CatalogColors.success:
-        return context.theme.colors.success;
+        return context.theme.customColors.success!;
       case _CatalogColors.info:
-        return context.theme.colors.info;
+        return context.theme.customColors.info!;
       case _CatalogColors.warning:
-        return context.theme.colors.warning;
+        return context.theme.customColors.warning!;
       case _CatalogColors.danger:
-        return context.theme.colors.danger;
+        return context.theme.customColors.danger!;
       case _CatalogColors.text:
-        return context.theme.colors.textColor;
+        return context.theme.customColors.textColor!;
     }
   }
 }
