+++ catalog/lib/catalog.dart
@@ -1,6 +1,6 @@
 library catalog;
 
-export 'theme/app_colors.dart';
+export 'theme/colors_extensions.dart';
 export 'theme/app_text_styles.dart';
 export 'theme/app_dimensions.dart';
 export 'theme/app_theme.dart';
